requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
brotli==1.2.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.6.10
crispy-bootstrap5==2025.6
cryptography==46.0.4
defusedxml==0.7.1
dj-database-url==2.3.0
dj-email-url==1.0.6
Django==6.0
django-allauth==65.13.1
django-cache-url==3.4.6
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
environs==11.2.1
gunicorn==23.0.0
idna==3.11
kombu==5.6.2
marshmallow==4.2.2
oauthlib==3.3.1
packaging==26.0
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python3-openid==3.2.0
redis==7.1.1
reportlab==4.4.9
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.5
stripe==14.1.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.9.0
